requirements.txt
pydub
opencv-python
numpy
pillow
moviepy